Finished up the FibonacciTree GUI so that it accepts a variable input value for the tree's level and also so it provides the user with some feedback when buttons are pressed for that GUI box. Changes: -Added a text field to the GUI for the user to enter a string version of an unsigned integer representing the tree's level. That value is parsed and used when building the tree now rather than a hard coded fixed value. -Added an error message string called "errorMsg" that has not only errors, but feedback for the user after they press a button. For example, if the user enters an invalid string into the Tree Level field, it displays a usage message at the bottom of the GUI box.
Notes:
-I considered changing the text field for the Tree Level to a slider, but since the example prompt took a string value and parsed it,
I figured I would do the same. If it weren't for that, I would have probably changed it to a slider that limited the range of possible levels
to something reasonable for a tree with this kind of memory footprint (maybe like 10 levels for this kind of demonstration) and also to impose a
suggested range of values upon the user. As is, the user could make a tree with 100000 levels, but it would take decades to generate that if
they even have enough RAM.